Rework reader loop and tasks
This changes the reader loop to stop on failure. If an assumption about
the data isn't true, it's better it stops immediately with a traceback
so the developer can figure out what's going on.

This changes the way celery tasks are used such that the script runs
celery in "always eager" mode. This enforces serialized execution of
synchronous tasks. Now when the script ends, all the work has been done.